server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "httpx>=0.28.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.41.0",
]
